Refactor DigitalTwin class and add DTAssets class
