+++ client/src/preview/route/digitaltwins/editor/sidebarFunctions.tsx
@@ -180,7 +180,7 @@ export const fetchAndSetFileContent = async (
   setFileType: Dispatch<SetStateAction<string>>,
 ) => {
   try {
-    const fileContent = await digitalTwin!.fileHandler.getFileContent(fileName);
+    const fileContent = await digitalTwin!.DTAssets.getFileContent(fileName);
     if (fileContent) {
       updateFileState(
         fileName,
